+++ packages/bpk-component-radio/src/BpkRadio-test.js
@@ -7,7 +7,7 @@ describe('BpkRadio', () => {
     const tree = renderer.create(
       <BpkRadio
         name="radio"
-        onChange={() => null}
+        label="Direct"
       />,
     ).toJSON();
     expect(tree).toMatchSnapshot();
@@ -18,7 +18,7 @@ describe('BpkRadio', () => {
       <BpkRadio
         id="radio"
         name="radio"
-        onChange={() => null}
+        label="Direct"
       />,
     ).toJSON();
     expect(tree).toMatchSnapshot();
@@ -28,19 +28,8 @@ describe('BpkRadio', () => {
     const tree = renderer.create(
       <BpkRadio
         name="radio"
-        onChange={() => null}
+        label="Direct"
         checked
-      />,
-    ).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-
-  it('should render correctly with label attribute', () => {
-    const tree = renderer.create(
-      <BpkRadio
-        name="radio"
-        onChange={() => null}
-        label="My label"
       />,
     ).toJSON();
     expect(tree).toMatchSnapshot();
@@ -50,7 +39,7 @@ describe('BpkRadio', () => {
     const tree = renderer.create(
       <BpkRadio
         name="radio"
-        onChange={() => null}
+        label="Direct"
         disabled
       />,
     ).toJSON();
@@ -61,8 +50,8 @@ describe('BpkRadio', () => {
     const tree = renderer.create(
       <BpkRadio
         name="radio"
+        label="Direct"
         value="my-value"
-        onChange={() => null}
       />,
     ).toJSON();
     expect(tree).toMatchSnapshot();